feat: organize src/core and src/chrome boundary
Move all Chrome-specific entry points (service worker, content script,
popup, options) and the Chrome MV3 manifest under `src/chrome/`. Create
an empty `src/core/` placeholder with a README documenting the
platform-agnostic boundary contract: no `chrome.*` / `browser.*`
imports, one-way dependency from `src/chrome/` to `src/core/`.

Rationale: feature parity with the Safari reference is the MVP bar, and
the engine, extraction, and overlay are intended to lift into a shared
package once the Chrome MVP ships. Establishing the directory boundary
before any non-trivial code lands keeps the eventual extraction cheap.

`manifest.ts` moves with the rest because it is Chrome-specific
(`chrome.runtime.ManifestV3` typing, MV3-shaped service worker config);
a future Safari port would have its own `src/safari/manifest.ts`. Icons
remain at the repo root since they are platform-agnostic assets.

Updates `CLAUDE.md` Repo orientation to reflect the actual layout.
STRUCTURE.md (issue #41) will expand this into a standalone reference
later.

Closes #9
