+++ src/chrome/background/index.ts
