+++ vite.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'vite';
 import { crx } from '@crxjs/vite-plugin';
-import manifest from './src/manifest';
+import manifest from './src/chrome/manifest';
 
 export default defineConfig({
   plugins: [
